+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added [telemetry 1.0](https://github.com/beam-telemetry/telemetry)
+- New telemetry events around queries
+
 ## [0.4.0] - 2021-06-10
 
 ### Added